Always restart updater systemd service on failures.
We saw a case where this was failing due to:
```
localhost systemd[1]: Dependency failed for flatcar-image-updater.service - Flatcar Image Updater Service.
localhost systemd[1]: flatcar-image-updater.service: Job flatcar-image-updater.service/start failed with result 'dependency'.
```
where the only dependency seems to be mounting a matchbox volume.
We can safely keep restarting this service on any failure (previously
Restart=on-failure should only be covering scenarios where the process
exits) and be gentle on intervals as we are not in a huge rush to fecth
new images after startup.